requirements.txt
boto3
python-magic
python-dotenv
typer